change int8 to int32

+++ w2w/w2w.py
@@ -211,7 +211,7 @@ def _replace_lcz_number(lcz, lcz_to_change):
          11, 12, 13, 14, 15, 16, 17]
     )
 
-    lcz_arr = lcz.data.flatten().astype(np.int8)
+    lcz_arr = lcz.data.flatten().astype(np.int32)
     df = pd.Series(lcz_arr, dtype=lcz_expected.dtype)
 
     d = dict(zip(lcz_to_change, lcz_expected))
@@ -219,7 +219,7 @@ def _replace_lcz_number(lcz, lcz_to_change):
     lcz_new = lcz.copy()
     lcz_new.data = df.map(d).values\
         .reshape(lcz.shape)\
-        .astype(np.int8)
+        .astype(np.int32)
 
     return lcz_new
 
@@ -284,7 +284,7 @@ def check_lcz_integrity(info: Dict[str, str], LCZ_BAND: int):
     if lcz.rio.crs != CRS.from_epsg(4326):
         lcz = lcz.rio.reproject("EPSG:4326")
         lcz.data = xr.where(lcz.data > 0, lcz.data, 0)
-        lcz.data = lcz.data.astype(np.int8)
+        lcz.data = lcz.data.astype(np.int32)
         print("> LCZ map reprojected to WGS84 (EPSG:4326).")
     else:
         print("> LCZ provided as WGS84 (EPSG:4326)")
@@ -293,7 +293,7 @@ def check_lcz_integrity(info: Dict[str, str], LCZ_BAND: int):
     _check_lcz_wrf_extent(lcz, wrf)
 
     # Write clean LCZ to file, used in all subsequent routines.
-    lcz.rio.to_raster(info['src_file_clean'], dtype=np.int8)
+    lcz.rio.to_raster(info['src_file_clean'], dtype=np.int32)
 
 
 def wrf_remove_urban(
@@ -476,7 +476,7 @@ def _ucp_resampler(
     )
 
     # Get LCZ class values only.
-    lcz_arr = src_data.data.astype(np.int8)
+    lcz_arr = src_data.data.astype(np.int32)
 
     # Set LCZ classes not in BUILT_LCZ to 0
     lcz_arr[~lcz_urb_mask] = 0
@@ -549,7 +549,7 @@ def _hgt_resampler(
     )
 
     # Get LCZ class values only.
-    lcz_arr = src_data.data.astype(np.int8)
+    lcz_arr = src_data.data.astype(np.int32)
 
     # Set LCZ classes not in BUILT_LCZ to 0
     lcz_arr[~lcz_urb_mask] = 0
@@ -723,7 +723,7 @@ def _hi_resampler(
     )
 
     # Get LCZ class values only.
-    lcz_arr = src_data.data.astype(np.int8)
+    lcz_arr = src_data.data.astype(np.int32)
 
     # Set LCZ classes not in BUILT_LCZ to 0
     lcz_arr[~lcz_urb_mask] = 0
